#17 [REFACTOR] 각 섹션 안에서 API 함수 호출하도록 수정
