+++ src/components/home/KoreanMoviesSection.tsx
@@ -1,10 +1,15 @@
-import MediaCardSection from "@/components/home/MediaCardSection";
+import MediaCardCarousel from "@/components/home/MediaCardCarousel";
 import { getKoreanMovies } from "@/lib/apis/tmdb";
 
 const KoreanMoviesSection = async () => {
   const data = await getKoreanMovies();
 
-  return <MediaCardSection title="Korean Movies" items={data.results} size="small" />;
+  return (
+    <div className="bg-black">
+      <p className="text-heading-2 pb-3 pl-1 text-white">Korean Movies</p>
+      <MediaCardCarousel items={data.results} size="small" />
+    </div>
+  );
 };
 
 export default KoreanMoviesSection;
