+++ src/components/home/NetflixOriginalsSection.tsx
@@ -1,10 +1,15 @@
-import MediaCardSection from "@/components/home/MediaCardSection";
+import MediaCardCarousel from "@/components/home/MediaCardCarousel";
 import { getNetflixOriginals } from "@/lib/apis/tmdb";
 
 const NetflixOriginalsSection = async () => {
   const data = await getNetflixOriginals();
 
-  return <MediaCardSection title="Netflix Originals" items={data.results} size="large" />;
+  return (
+    <div className="bg-black">
+      <p className="text-heading-2 pb-3 pl-1 text-white">Netflix Originals</p>
+      <MediaCardCarousel items={data.results} size="large" />
+    </div>
+  );
 };
 
 export default NetflixOriginalsSection;
